pyproject.toml
[project]
name = "flowbitai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio-redis>=0.16.0",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.62",
    "langchain-google-genai>=2.1.5",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
    "streamlit>=1.45.1",
]
